[Bug 26333 and  26080, 27100 ]. Quick fix for the test names being randomly generated from testkit.
"id" in the root testkit.cfg can be used to generate a custom id. The id format should use the same id format used in pbreports.

If the id is not in the cfg, the job id will be set from the basename of the dirname.

The cfg spec should be updated to REQUIRE the job_id, description and author to be present.