Redo logging config in pbsmrtpipe
Added new pbsmrtpipe.options.debug_mode that can specified in the preset.xml

There's still some need fixes, but this is good enough for now. The fundamentally model should be pushed down to pbcommand.

Any tools called via python -m (used in the RTC driver) are not going to have the log setup correctly because of disable_existing_loggers is not set correctly.