Use file-tailing, not pipe
This is an even better, simpler way to get the stack-trace: Just send
the subproc output directly to files on disk. We can periodically
check for new content, like `tail -f`, and log it. With NFS, we might
sometimes not show everyone, but that's not an error. The important
thing is that the output is *always* in the files, so having a
stack-trace on Ctrl-C is trivial.