Dropped defaults
I dropped the defaults here:
* https://github.com/PacificBiosciences/FALCON-pbsmrtpipe/commit/32e4dd3a728c84c429c42360f278eafa4091c4a2#diff-9a50acc93f01a2ea3ec651d346ee8b4bL35

I am not sure whether the TC needs to updated immediately, but needs
to be updated eventually. Keep an eye on this:
* http://jenkins:8080/job/Ubuntu1404_Mainline_SA3_HGAP_4_tests/

Without the defaults in `@registry`, the TC registrar does
not emit anything. I have no idea whether that is desired.
However, I do need to be able to update the TCs in an automated
way.