DEP-435: print cluster_cmd and underlying jms submit command to master.log
Print cluster_cmd and underlying jms submit command to master.log.  For
printing the underlying jms submit command, we basically detect if we are
submitting to the cluster with runjmscmd.  If so, we will first call the
command with the --printcmd option and print the stdout to the log file (if
the command completes successfully).

So two lines will be added to the master.log in production installs, the
full runjmscmd that is being executed, and the underlying JMS command that is
being called by runjmscmd.